Make ExceptionAttempt and ResultAttempt public
